% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{fnames}
\alias{fnames}
\title{Display formatted names}
\usage{
fnames(x, flags = "")
}
\arguments{
\item{x}{An object that can have names. Generally a \code{data.frame}.}

\item{flags}{A character scalar (default is the empty string). Add a \code{+} (resp \code{-}) to
sort the names in alphabetical order (\code{-}: decreasing). Then you can add a \verb{#} to request
the display of the position of the variables. You must respect this order.
Finally the rest is a \emph{quickselect} pattern
which can be used to select only subsets of variables. See the help of [\link{selvars}] to get
more information on how \emph{quickselect} works.}
}
\value{
This function is for display only and does not return anything.
}
\description{
Simple function to display the names of a data.frame, or anything with names,
in a human readable way. You can control the formatting with the flags.
}
\details{
If the position of the variables is requested but the variables are not in sequential
positions (due to sorting or subselecting), then the position is displayed in parentheses
just after the variable name. Otherwise, it is displayed on the left.
}
\examples{

# display the variables of a data set
fnames(iris)

# data set with many names
base_many = t(mtcars)
fnames(base_many)

# adding the position
fnames(base_many, "#")

# selecting only a subset: the ones containing two digits
fnames(base_many, "#@\\\\d\\\\d")

# same but selecting volvo + mazda
fnames(base_many, "-#^vol, ^maz")

}
