% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smagick_main.R
\name{print.smagick}
\alias{print.smagick}
\title{Simple print function for objects of class \code{smagick}}
\usage{
\method{print}{smagick}(x, ...)
}
\description{
Print \code{smagick} character vectors in a nice way.
}
\details{
x A \code{smagick} object, obtained from the function \code{smagick}.
... Not currently used.
}
\examples{

cars = row.names(mtcars)

print(cars)

smagick("/The first cars are:, {6 first ? cars}")


}
\seealso{
A few basic operation: \code{\link[=str_is]{str_is()}}, \code{\link[=str_get]{str_get()}}, \code{\link[=str_clean]{str_clean()}}. Chain basic operations with \code{\link[=str_ops]{str_ops()}}.
String interpolation combined with operation chaining: \code{\link[=smagick]{smagick()}}.
}
\author{
Laurent Berge
}
