% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smagic_main.R
\name{setSmagic}
\alias{setSmagic}
\title{Set defaults for smagic}
\usage{
setSmagic(
  .class = "smagic",
  .delim = c("{", "}"),
  .sep = "",
  .data.table = TRUE,
  reset = FALSE
)
}
\arguments{
\item{.class}{Character vector representing the class to give to the object returned.
By default it is \code{NULL}. Note that the class \code{smagic} has a specific \code{print} method, usually
nicer for small vectors (it \code{\link[base:cat]{base::cat()}}s the elements).}

\item{.delim}{Character vector of length 1 or 2. Default is \code{c("{", "}")}. Defines
the opening and the closing delimiters for interpolation.

If of length 1, it must be of the form: 1) the opening delimiter,
2) a single space, 3) the closing delimiter. Ex: \code{".[ ]"} is equivalent to \code{c(".[", "]")}.
The default value is equivalent to \code{"{ }"}.

[ ]: R:\%20
[", "]: R:\%22,\%20\%22}

\item{.sep}{Character scalar, default is the empty string \code{""}. It is used to collapse all
the elements in \code{...} before applying any operation.}

\item{.data.table}{Logical, default is \code{TRUE}. If you use \code{smagic} in a \code{data.table} call
and interpolate variables within the \code{data.table}, you want this argument to be \code{TRUE}.
It only incurs a small overhead.}

\item{reset}{Logical scalar, default is \code{FALSE}. Whether to reset all values.}
}
\description{
Se the default values for a few arguments of the function \code{\link[=smagic]{smagic()}}.
}
\details{
By default, each call to \code{setSmagic} adds modifications to the default values.
To set a few default values and resetting the others, you need to use \code{reset = TRUE}.

If you are a package developer: note that these options are scoped. This means that if you want to
change the behavior of \code{smagic} within your package: you can and there will
be no conflic if the user also uses smagic with her own, different, options.
To define options, you should add a call to \code{setSmagic} in your \code{.onLoad()} function
which is run at package startup (see \code{help(".onLoad")}).
It is important that the \code{setSmagic} call should be in that function.
}
\examples{

# we change the default display of the results of smagic
setSmagic(.class = "smagic")
smagic("{S!x, y}{2 each?1:2}")

# back to a regular character vector
setSmagic(reset = TRUE)
smagic("{S!x, y}{2 each?1:2}")

}
\author{
Laurent Berge
}
