% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smagic_main.R
\name{smagic_alias}
\alias{smagic_alias}
\title{Create \code{smagic} aliases with custom defaults}
\usage{
smagic_alias(
  .sep = "",
  .vectorize = FALSE,
  .delim = c("{", "}"),
  .last = NULL,
  .collapse = NULL,
  .check = TRUE,
  .class = NULL,
  .data.table = TRUE,
  .namespace = NULL
)
}
\arguments{
\item{.sep}{Character scalar, default is the empty string \code{""}. It is used to collapse all
the elements in \code{...} before applying any operation.}

\item{.vectorize}{Logical scalar, default is \code{FALSE}. If \code{TRUE}, Further, elements in \code{...} are
NOT collapsed together, but instead vectorised.}

\item{.delim}{Character vector of length 1 or 2. Default is \code{c("{", "}")}. Defines
the opening and the closing delimiters for interpolation.

If of length 1, it must be of the form: 1) the opening delimiter,
2) a single space, 3) the closing delimiter. Ex: \code{".[ ]"} is equivalent to \code{c(".[", "]")}.
The default value is equivalent to \code{"{ }"}.

[ ]: R:\%20
[", "]: R:\%22,\%20\%22}

\item{.last}{Character scalar, a function, or \code{NULL} (default). If provided and character:
it must be an \code{smagic} chain of operations of the form \code{"'arg1'op1, op2, etc"}. All these operations
are applied just before returning the vector. If a function,
it will be applied to the resulting vector.}

\item{.collapse}{Character scalar, default is \code{NULL}. If provided, the character vector
that should be returned is collapsed with the value of this argument. This leads
to return a string of length 1.}

\item{.check}{Logical scalar, default is \code{TRUE}. Whether to enable error-handling (i.e.
human readable error messages).
Without error-handling you can save something of the order of 40us. Useful only
in long loops.}

\item{.class}{Character vector representing the class to give to the object returned.
By default it is \code{NULL}. Note that the class \code{smagic} has a specific \code{print} method, usually
nicer for small vectors (it \code{\link[base:cat]{base::cat()}}s the elements).}

\item{.data.table}{Logical, default is \code{TRUE}. If you use \code{smagic} in a \code{data.table} call
and interpolate variables within the \code{data.table}, you want this argument to be \code{TRUE}.
It only incurs a small overhead.}
}
\description{
Utility to easily create \code{smagic} aliases with custom default
}
\details{
Use this function if you want to change \code{smagic} default values. For example,
if you want the interpolation to be done with "{{}}" (instead of \code{{}}) or if you want the
default separation to be the space (instead of the empty string).
}
\examples{

# we create the function sma2 with different defaults
sma2 = smagic_alias(.delim = ".[ ]", .sep = "", .class = `smagick`)

person = "john doe"
sma2("Hello", "{title ? person}")



}
\author{
Laurent Berge
}
