% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smagic_main.R
\name{smagic_register}
\alias{smagic_register}
\title{Register custom operations to apply them in smagic}
\usage{
smagic_register(fun, alias, valid_options = NULL, scope = NULL)
}
\arguments{
\item{fun}{A function which must have at least the arguments 'x' and '...'.
Additionnaly, it can have the arguments: 'argument', 'options', 'group', 'group_flag'.
This function must return a vector.
This function will be internally called by \code{smagic} in the form
\code{fun(x, argument, options, group, group_flag)}.\code{x}: the value to which the
operation applies. \code{argument}: the quoted \code{smagic} argument (always character).
\code{options}: a character vector of \code{smagic} options. The two last arguments are of use
only in group-wise operations if \code{fun} changes the lengths of vectors. \code{group}: an index of
the group to which belongs each observation (integer). \code{group_flag}: value between 0
and 2; 0: no grouping operation requested; 1: keep track of groups; 2: apply grouping.}

\item{alias}{Character scalar, the name of the operation.}

\item{valid_options}{A character vector or NULL (default). Represents a list of
valid options for the operation. This is used: a) to enable auto-completion,
b) for error-handling purposes.}
}
\description{
Extends the capabilities of \code{\link[=smagic]{smagic()}} by adding any custom operation
}
\examples{

# let's create an operation that adds markdown emphasis
# to elements of a vector

# A) define the function
fun_emph = function(x, ...) paste0("*", x, "*")

# B) register it
smagic_register(fun_emph, "emph")

# C) use it
x = str_vec("right, now")
smagic("Take heed, {emph, c? x}.")

#
# now let's add the option "strong"
fun_emph = function(x, options, ...) {
  if("strong" \%in\% options){
    paste0("***", x, "***")
  } else {
    paste0("*", x, "*")
  }
}

smagic_register(fun_emph, "emph", "strong")

x = str_vec("right, now")
smagic("Take heed, {emph.strong, c? x}.")

#
# now let's add an argument
fun_emph = function(x, argument, options, ...){
  arg = argument
  if(nchar(arg) == 0) arg = "*"
  
  if("strong" \%in\% options){
    arg = paste0(rep(arg, 3), collapse = "")
  }
  
  paste0(arg, x, arg)
}

smagic_register(fun_emph, "emph", "strong")

x = str_vec("right, now")
smagic("Take heed, {'_'emph.s, c? x}.")




}
\seealso{
\code{smagic()}
}
\author{
Laurent R. Berge
}
