% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{str_fill}
\alias{str_fill}
\title{Fills a character string up to a size}
\usage{
str_fill(
  x = "",
  n = NULL,
  symbol = " ",
  right = FALSE,
  center = FALSE,
  na = "NA"
)
}
\arguments{
\item{x}{A character vector.}

\item{n}{Integer scalar, possibly equal to \code{NULL} (default). The size up to which the character
vector will be filled. If \code{NULL} (default), it is set to the largest width in the character vector \code{x}.
To handle how the character is filled, see the arguments \code{symbol}, \code{right} and \code{center}.}

\item{symbol}{Character scalar of length 1, default is a space (" "). It is the symbol with which
the string will be filled.}

\item{right}{Logical scalar, default is \code{FALSE}. If \code{TRUE}, then the filling of the string is
done from the left, leading to right-alignment.}

\item{center}{Logical scalar, default is \code{FALSE}. If \code{TRUE}, then the filling of the string will
be balanced so as to center the strings.}

\item{na}{Character scalar or \code{NA}. Default is "NA" (a character string!). What happens to NAs: by default
they are replaced by the character string "NA".}
}
\description{
Fills a character string up to a size and handles multibyte encodings
(differently from sprintf).
}
\details{
If you use character filling of the form \code{sprintf("\% 20s", x)} with \verb{x``containing multibyte characters, you may be suprised that all character strings do not end up at the same lenght (the occurrence of this problem depends on many things: encodings are a mess). }str_fill`
uses only base R functions to compensate this. It is slightly slower but, in general, safer.

It also looks a bit like \code{\link[base:format]{base::format()}}, but slightly different (and a bit faster, but more restrictive).
}
\examples{

x = c("apple", "pineapple") 

# simple fill with blank
cat(paste0(str_fill(x), ":", c(3, 7), "€"), sep = "\n")

# center fill
cat(paste0(str_fill(x, center = TRUE), ":", c(3, 7), "€"), sep = "\n")

# changing the length of the fill and the symbol used for filling
cat(paste0(str_fill(x), ":", str_fill(c(3, 7), 3, "0", right = TRUE), "€"), sep = "\n")

# na behavior: default/NA/other
x = c("hello", NA) 
str_fill(x)
str_fill(x, na = NA)
str_fill(x, na = "(missing)")


}
\seealso{
A few basic operation: \code{\link[=str_is]{str_is()}}, \code{\link[=str_get]{str_get()}}, \code{\link[=str_clean]{str_clean()}}. Chain basic operations with \code{\link[=str_ops]{str_ops()}}.
String interpolation combined with operation chaining: \code{\link[=smagick]{smagick()}}.
}
\author{
Laurent R. Berge
}
