% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{str_op}
\alias{str_op}
\title{Chains basic operations to character vectors}
\usage{
str_op(x, op, pre_unik = NULL)
}
\arguments{
\item{x}{A character vector. If not a character vector but atomistic (i.e. not a list),
it will be converted to a character vector.}

\item{op}{Character \strong{scalar}. Character scalar containing the comma separated values
of operations to perform to the vector. For help type \code{cub("--help")}.}

\item{pre_unik}{Logical scalar, default is \code{NULL}. Whether to first unique the vector
before applying the possibly costly string operations, and merging back the result.
For very large vectors with repeated values the time gained can be substantial. By
default, this is \code{TRUE} for vector of length 1M or more.}
}
\value{
In general it returns a character vector. It may ba of a length different from the original
one, depending on the operations performed.
}
\description{
Simple tool to perform multiple operations to character vectors.
}
\author{
Laurent Berge
}
