% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{_index_}
\alias{_index_}
\alias{stringmagick}
\alias{stringmagick-package}
\title{String operations and interpolation: Magick edition}
\description{
The stringmagick package introduces a new language tailored to create complex character strings.
Use the function smagick to interpolate a variable inside a character string and to apply,
on-the-fly, any arbitrary string operation.
}
\details{
It supports over 50 basic operations, seamless customization, nesting,
pluralization, and much more. Writing complex character strings has never been that easy!

To get an overview of the package: \url{take the tour}. To get started with smagick,
see \url{the vignette}.

This package also enhances regular string functions (like grep and co) with
specialized functions to:
\itemize{
\item detect combinations of regular expressions
\item chain basic string operations
\item clean character string vectors
}

See more in the \url{dedicated vignette}.
}
\keyword{internal}
