% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smagic_main.R
\name{print.smagic}
\alias{print.smagic}
\title{Simple print function for objects of class \code{smagic}}
\usage{
\method{print}{smagic}(x, ...)
}
\arguments{
\item{x}{A \code{smagic} object, obtained from the function \code{smagic}.}

\item{...}{Not currently used.}
}
\description{
Print \code{smagic} character vectors in a nice way.
}
\examples{

cars = row.names(mtcars)

print(cars)

str_vec("The first cars are:, {6 first ? cars}")


}
\seealso{
A few basic operation: \code{\link[=str_is]{str_is()}}, \code{\link[=str_get]{str_get()}}, \code{\link[=str_clean]{str_clean()}}. Chain basic operations with \code{\link[=str_ops]{str_ops()}}.
Use \code{\link[=str_vec]{str_vec()}} to create simple string vectors.
String interpolation combined with operation chaining: \code{\link[=smagic]{smagic()}}. You can change \code{smagic}
default values with \code{\link[=smagic_alias]{smagic_alias()}} and add custom operations with \code{\link[=smagic_register_fun]{smagic_register_fun()}}.
}
\author{
Laurent Berge
}
